Updated Login Process
Current Setup is designed to connect to a database called "Library" on
MySQL for testing. It then uses the employeeinformation table which
contains "FirstName", "LastName", "ID", "Username", "Password" and
"AdminState". Using this information it welcomes the user to the system
and provides the link the normal logged in Index page. Also added a
logout script page.